+++ pallets/bridge/src/lib.rs
@@ -76,13 +76,13 @@ decl_event!(
         /// Minted
         Minted(AccountId, u32, AccountId, Balance),
         /// Minting Request added to the queue
-        MintQueued(AccountId, u32, AccountId, Balance, Vec<u8>),
+        MintQueued(AccountId, u32, AccountId, Balance, Vec<u8>, Vec<u8>),
         /// Already minted the same transaction
         AlreadyMinted(AccountId, u32, AccountId, Balance),
         /// Burned
         Burned(AccountId, u32, AccountId, Balance),
         /// Burning Request added to the queue
-        BurnQueued(AccountId, u32, AccountId, Balance, Vec<u8>),
+        BurnQueued(AccountId, u32, AccountId, Balance, Vec<u8>, Vec<u8>),
         /// Already burned the same transaction
         AlreadyBurned(AccountId, u32, AccountId, Balance),
     }
@@ -412,7 +412,7 @@ decl_module! {
             }
  
             // update the counter for the minting requests of the transaction
-            let mut key = token;
+            let mut key = token.clone();
             key.push(b'-');
             key.append(&mut recipient.encode());
             key.push(b'-');
@@ -425,7 +425,7 @@ decl_module! {
             let nmr=MintCounter::get(&key);
             // thresold not reached
             match nmr.cmp(&internalthreshold) {
-                Ordering::Less => Self::deposit_event(RawEvent::MintQueued(signer, asset_id, recipient, amount, transaction_id.clone())),
+                Ordering::Less => Self::deposit_event(RawEvent::MintQueued(signer, asset_id, recipient, amount, transaction_id.clone(), token.clone())),
                 Ordering::Greater => Self::deposit_event(RawEvent::AlreadyMinted(signer, asset_id, recipient, amount)),
                 Ordering::Equal => {
                     // check it's not already confirmed
@@ -494,7 +494,7 @@ decl_module! {
             }
 
             // update the counter for the minting requests of the transaction
-            let mut key = token;
+            let mut key = token.clone();
             key.push(b'-');
             key.append(&mut recipient.encode());
             key.push(b'-');
@@ -507,7 +507,7 @@ decl_module! {
             // get the number of burning requests
             let nmr=BurnCounter::get(&key);
             match nmr.cmp(&internalthreshold) {
-                Ordering::Less => Self::deposit_event(RawEvent::BurnQueued(signer, asset_id, recipient, amount, transaction_id.clone())),
+                Ordering::Less => Self::deposit_event(RawEvent::BurnQueued(signer, asset_id, recipient, amount, transaction_id.clone(), token.clone())),
                 Ordering::Greater => Self::deposit_event(RawEvent::AlreadyBurned(signer, asset_id, recipient, amount)),
                 Ordering::Equal => {
                     // check it's not already confirmed
@@ -826,36 +826,6 @@ impl<T: Config> Module<T> {
             lb= b;
         }
         result
-    }
-
-    pub fn mint_tracker_contains(transaction_id: Vec<u8>, signer: T::AccountId) -> bool {
-        TransactionMintTracker::<T>::contains_key(transaction_id,&signer)
-    }
-
-    pub fn burn_tracker_contains(transaction_id: Vec<u8>, signer: T::AccountId) -> bool {
-        TransactionBurnTracker::<T>::contains_key(transaction_id.clone(),&signer)
-    }
-
-    pub fn is_threshold_mint(token:Vec<u8>) -> bool {
-        if let Some(content) = Settings::get(&token) {
-            let internalthresholdv = Self::json_get_value(content.clone(),"internalthreshold".as_bytes().to_vec());
-            let internalthreshold = vecu8_to_u32(internalthresholdv);
-            let nmr= MintCounter::get(&token);
-            nmr >= internalthreshold
-        } else {
-            false
-        }
-    }
-
-    pub fn is_threshold_burn(token:Vec<u8>) -> bool {
-        if let Some(content) = Settings::get(&token) {
-            let internalthresholdv = Self::json_get_value(content.clone(),"internalthreshold".as_bytes().to_vec());
-            let internalthreshold = vecu8_to_u32(internalthresholdv);
-            let nmr= BurnCounter::get(&token);
-            nmr >= internalthreshold
-        } else {
-            false
-        }
     }
 
 }
